Fixed: Issue where parachutes were not added in dedicated environment
Not going to lie, it's 0155 and I have no idea why this wasn't working in dedi, by all accounts it should.

Yeeting the commands to swap backpack and add a parachute to the clients via a remoteExec disguised as a CBA function seems to have done the trick for now.